Match-id-169b47a323cb0326d90a5e0e44b9c97eb02de747

+++ test/run_ut.sh
@@ -5,6 +5,8 @@ CUR_DIR=$(dirname $(readlink -f $0))
 TOP_DIR=$(readlink -f ${CUR_DIR}/..)
 TEST_DIR=${TOP_DIR}/"test"
 SRC_DIR=${TOP_DIR}/"src"
+TEST_LANG=$1
+ENABLE_CACHE=$2
 ret=0
 
 clean() {
@@ -35,15 +37,44 @@ function fn_build_googletest()
   fi
 }
 
+function fn_build_mock_cpp()
+{
+  cd ${CUR_DIR}/../opensource
+  MOCK_CPP_DIR="${CUR_DIR}/../opensource/mock_cpp"
+  if [ ! -d "$MOCK_CPP_DIR" ]; then
+    git clone https://szv-y.codehub.huawei.com/d00437232/mock_cpp.git mock_cpp -b msprof
+  else
+      echo "opensource/mock_cpp already exists. no need to download."
+  fi
+
+  if [ ! -d "$MOCK_CPP_DIR/lib" ]; then
+    cd mock_cpp
+    mkdir build
+    cd build
+    cmake -DCMAKE_INSTALL_PREFIX=$MOCK_CPP_DIR/mockcpp -DMOCKCPP_XUNIT=gtest \
+      -DMOCKCPP_XUNIT_HOME=${CUR_DIR}/../opensource/googletest ..
+    make -j20
+    make install
+  fi
+}
+
 run_test_cpp() {
   cd ${TEST_DIR}/..
-  bash build.sh
+
+  if [ "$ENABLE_CACHE" != "1" ]; then
+    bash build.sh
+  fi
   if [ $? -ne 0 ]; then
     echo "Build ms_service_profiler failed"
     exit 1
   fi
   cd ${TEST_DIR}
-  mkdir -p test_build && cd test_build && rm * -rf && cmake ../ut/cpp_test && make -j 4
+
+  if [ "$ENABLE_CACHE" != "1" ]; then
+    mkdir -p test_build && cd test_build && rm * -rf && cmake ../ut/cpp_test && make -j 4
+  else
+    mkdir -p test_build && cd test_build && cmake ../ut/cpp_test && make -j 4
+  fi
   if [ $? -ne 0 ]; then
     echo "Build test failed"
     exit 1
@@ -63,18 +94,18 @@ run_test_cpp() {
   lcov_opt="--rc lcov_branch_coverage=1 --rc geninfo_no_exception_branch=1"
   lcov -c -d ./test_build/CMakeFiles/st_server_profiler.dir -o ./coverage/st_server_profiler.info -b ./coverage $lcov_opt
   lcov -c -d ./test_build/CMakeFiles/ut_server_profiler.dir -o ./coverage/ut_server_profiler.info -b ./coverage $lcov_opt
-  lcov -a ./coverage/ut_server_profiler.info  -a ./coverage/st_server_profiler.info  -o ./coverage/test_server_profiler.info
+  lcov -a ./coverage/ut_server_profiler.info  -a ./coverage/st_server_profiler.info  -o ./coverage/test_server_profiler.info  $lcov_opt
 
-  lcov -r ./coverage/test_server_profiler.info '*platform*' -o ./coverage/test_server_profiler.info $lcov_opt
-  lcov -r ./coverage/test_server_profiler.info '*opensource*' -o ./coverage/test_server_profiler.info $lcov_opt
-  lcov -r ./coverage/test_server_profiler.info '*test*' -o ./coverage/test_server_profiler.info $lcov_opt
-  lcov -r ./coverage/test_server_profiler.info '*c++*' -o ./coverage/test_server_profiler.info $lcov_opt
-  lcov -r ./coverage/test_server_profiler.info '/usr/include/*' -o ./coverage/test_server_profiler.info $lcov_opt
-  lcov -r ./coverage/test_server_profiler.info '*nlohmann*' -o ./coverage/test_server_profiler.info $lcov_opt
+  lcov -r ./coverage/test_server_profiler.info '*platform*' -o ./coverage/test_server_profiler.info $lcov_opt -q
+  lcov -r ./coverage/test_server_profiler.info '*opensource*' -o ./coverage/test_server_profiler.info $lcov_opt -q
+  lcov -r ./coverage/test_server_profiler.info '*test*' -o ./coverage/test_server_profiler.info $lcov_opt -q
+  lcov -r ./coverage/test_server_profiler.info '*c++*' -o ./coverage/test_server_profiler.info $lcov_opt -q
+  lcov -r ./coverage/test_server_profiler.info '/usr/include/*' -o ./coverage/test_server_profiler.info $lcov_opt -q
+  lcov -r ./coverage/test_server_profiler.info '*nlohmann*' -o ./coverage/test_server_profiler.info $lcov_opt -q
 
   genhtml ./coverage/test_server_profiler.info -o ./coverage/report --branch-coverage
   cd coverage
-  tar -zcvf report.tar.gz ./report
+  tar -zcvf report.tar.gz ./report  -q
   echo show report using cmd: python -m http.server -d ./coverage/report
 }
 
@@ -94,14 +125,20 @@ run_test_python() {
 }
 
 run_test() {
-  run_test_cpp
-  run_test_python
+  if [ "$TEST_LANG" != "python" ]; then
+    run_test_cpp
+  fi
+
+  if [ "$TEST_LANG" != "cpp" ]; then
+    run_test_python
+  fi
 }
 
 main() {
   export VERBOSE=1
   cd ${TEST_DIR}
   fn_build_googletest
+  fn_build_mock_cpp
   clean
   run_test
   echo "UT Success"
